+++ tests/test_visitors/test_ast/test_general/test_wrong_name/test_function_names.py
@@ -10,7 +10,7 @@ from wemake_python_styleguide.violations.naming import (
     WrongVariableNameViolation,
 )
 from wemake_python_styleguide.visitors.ast.naming import (
-    BAD_VARIABLE_NAMES,
+    VARIABLE_NAMES_BLACKLIST,
     WrongNameVisitor,
 )
 
@@ -28,7 +28,7 @@ class Input:
 """
 
 
-@pytest.mark.parametrize('bad_name', BAD_VARIABLE_NAMES)
+@pytest.mark.parametrize('bad_name', VARIABLE_NAMES_BLACKLIST)
 @pytest.mark.parametrize('code', [
     function_name,
     async_function_name,